refactor(align-deps): migrate vigilant mode
