+++ packages/align-deps/src/index.ts
@@ -1,20 +1,19 @@
 export { capabilitiesFor } from "./capabilities";
-export { checkPackageManifest } from "./check";
 export { cli } from "./cli";
+export { checkPackageManifest } from "./commands/check";
+export { checkPackageManifestUnconfigured } from "./commands/vigilant";
 export { updatePackageManifest } from "./manifest";
-
 export type {
   Args,
   CapabilitiesOptions,
-  CheckOptions,
   Command,
   DependencyType,
   ExcludedPackage,
   ManifestProfile,
   MetaPackage,
+  Options,
   Package,
   Profile,
   ProfilesInfo,
   ProfileVersion,
-  VigilantOptions,
 } from "./types";
