+++ packages/align-deps/test/__fixtures__/config-custom-profiles-only/packageSpecificProfiles.js
@@ -4,6 +4,14 @@ module.exports = {
       name: "react-native",
       version: "0.64.3",
     },
+    "core-android": {
+      name: "react-native",
+      version: "0.64.3",
+    },
+    "core-ios": {
+      name: "react-native",
+      version: "0.64.3",
+    },
     react: {
       name: "react",
       version: "17.0.2",
@@ -11,6 +19,14 @@ module.exports = {
   },
   0.65: {
     core: {
+      name: "react-native",
+      version: "0.65.2",
+    },
+    "core-android": {
+      name: "react-native",
+      version: "0.65.2",
+    },
+    "core-ios": {
       name: "react-native",
       version: "0.65.2",
     },
