+++ packages/align-deps/test/check.test.ts
@@ -1,10 +1,8 @@
-import semverCoerce from "semver/functions/coerce";
 import {
   checkPackageManifest,
   containsValidPresets,
   containsValidRequirements,
-  getCheckConfig,
-} from "../src/check";
+} from "../src/commands/check";
 import profile_0_68 from "../src/presets/microsoft/react-native/profile-0.68";
 import profile_0_69 from "../src/presets/microsoft/react-native/profile-0.69";
 import profile_0_70 from "../src/presets/microsoft/react-native/profile-0.70";
@@ -596,81 +594,5 @@ describe("checkPackageManifest({ kitType: 'library' }) (backwards compatibility)
     expect(checkPackageManifest("package.json", defaultOptions)).toBe(
       "success"
     );
-  });
-});
-
-describe("getCheckConfig", () => {
-  const rnxKitConfig = require("@rnx-kit/config");
-  const fs = require("fs");
-
-  const mockManifest = {
-    name: "@rnx-kit/align-deps",
-    version: "0.0.1",
-  };
-
-  const defaultOptions = { loose: false, write: false };
-
-  beforeEach(() => {
-    fs.__setMockContent(mockManifest);
-    rnxKitConfig.__setMockConfig();
-  });
-
-  afterAll(() => {
-    jest.clearAllMocks();
-  });
-
-  test("returns early if the package is unconfigured", () => {
-    expect(getCheckConfig("package.json", defaultOptions)).toBe(0);
-  });
-
-  test("returns default values given react-native version", () => {
-    const reactNativeVersion = "^0.64";
-    rnxKitConfig.__setMockConfig({ reactNativeVersion });
-
-    expect(getCheckConfig("package.json", defaultOptions)).toEqual({
-      capabilities: [],
-      kitType: "library",
-      manifest: mockManifest,
-      reactNativeVersion,
-      reactNativeDevVersion: semverCoerce(reactNativeVersion)?.version,
-    });
-  });
-
-  test("uses react-native version provided by vigilant flag if unspecified", () => {
-    const reactNativeVersion = "^0.64";
-    rnxKitConfig.__setMockConfig({ customProfiles: "" });
-
-    expect(
-      getCheckConfig("package.json", {
-        ...defaultOptions,
-        supportedVersions: reactNativeVersion,
-        targetVersion: reactNativeVersion,
-      })
-    ).toEqual({
-      capabilities: [],
-      kitType: "library",
-      manifest: mockManifest,
-      reactNativeVersion,
-      reactNativeDevVersion: reactNativeVersion,
-    });
-  });
-
-  test("does not overwrite existing config with the version provided by vigilant flag", () => {
-    const reactNativeVersion = "^0.64";
-    rnxKitConfig.__setMockConfig({ reactNativeVersion });
-
-    expect(
-      getCheckConfig("package.json", {
-        ...defaultOptions,
-        supportedVersions: "1000.0",
-        targetVersion: "1000.0",
-      })
-    ).toEqual({
-      capabilities: [],
-      kitType: "library",
-      manifest: mockManifest,
-      reactNativeVersion,
-      reactNativeDevVersion: semverCoerce(reactNativeVersion)?.version,
-    });
   });
 });
