+++ packages/align-deps/test/dependencies.test.ts
@@ -1,10 +1,6 @@
 import { PackageManifest, readPackage } from "@rnx-kit/tools-node/package";
 import path from "path";
-import {
-  gatherRequirements,
-  getRequirements,
-  visitDependencies,
-} from "../src/dependencies";
+import { gatherRequirements, visitDependencies } from "../src/dependencies";
 import profile_0_63 from "../src/presets/microsoft/react-native/profile-0.63";
 import profile_0_64 from "../src/presets/microsoft/react-native/profile-0.64";
 
@@ -223,141 +219,6 @@ describe("gatherRequirements()", () => {
     expect(consoleWarnSpy).toBeCalledWith(
       "warn",
       expect.stringContaining("No profiles could satisfy all requirements")
-    );
-  });
-});
-
-describe("getRequirements()", () => {
-  const consoleErrorSpy = jest.spyOn(global.console, "error");
-  const consoleWarnSpy = jest.spyOn(global.console, "warn");
-  const defaultOptions = { loose: false };
-
-  afterEach(() => {
-    consoleErrorSpy.mockReset();
-    consoleWarnSpy.mockReset();
-  });
-
-  test("gets requirements from all dependencies", () => {
-    const [fixture, manifest] = useFixture("awesome-repo");
-    const { reactNativeVersion, capabilities } = getRequirements(
-      "^0.63 || ^0.64",
-      "app",
-      manifest,
-      fixture,
-      undefined,
-      defaultOptions
-    );
-
-    expect(reactNativeVersion).toBe("^0.63 || ^0.64");
-
-    expect(capabilities.sort()).toEqual([
-      "animation",
-      "netinfo",
-      "storage",
-      "webview",
-    ]);
-
-    expect(consoleErrorSpy).not.toBeCalled();
-    expect(consoleWarnSpy).not.toBeCalled();
-  });
-
-  test("gets requirements from all dependencies with custom profiles", () => {
-    const cyberdyne = { name: "cyberdyne", version: "1.0.0", devOnly: true };
-    const skynet = { name: "skynet", version: "1.0.0" };
-    jest.mock(
-      "awesome-align-deps-preset",
-      () => ({
-        "0.63": {
-          [cyberdyne.name]: cyberdyne,
-          [skynet.name]: skynet,
-        },
-        "0.64": {
-          [cyberdyne.name]: cyberdyne,
-          [skynet.name]: skynet,
-        },
-      }),
-      { virtual: true }
-    );
-
-    const [fixture, manifest] = useFixture("awesome-repo-extended");
-    const { reactNativeVersion, capabilities } = getRequirements(
-      "^0.63 || ^0.64",
-      "app",
-      manifest,
-      fixture,
-      "awesome-align-deps-preset",
-      defaultOptions
-    );
-
-    expect(reactNativeVersion).toBe("^0.63 || ^0.64");
-
-    expect(capabilities.sort()).toEqual([
-      "animation",
-      "netinfo",
-      "skynet",
-      "storage",
-      "webview",
-    ]);
-
-    expect(consoleErrorSpy).not.toBeCalled();
-    expect(consoleWarnSpy).not.toBeCalled();
-  });
-
-  test("throws if no profiles can satisfy required React Native version", () => {
-    expect(() =>
-      getRequirements(
-        "0.60.6",
-        "app",
-        {
-          name: "@rnx-kit/align-deps",
-          version: "1.0.0",
-        },
-        "",
-        undefined,
-        defaultOptions
-      )
-    ).toThrow();
-
-    expect(consoleErrorSpy).not.toBeCalled();
-    expect(consoleWarnSpy).not.toBeCalled();
-  });
-
-  test("throws if no profiles can satisfy requirement of dependencies", () => {
-    const [fixture, manifest] = useFixture("no-profile-satisfying-deps");
-    expect(() =>
-      getRequirements(
-        "^0.64",
-        "app",
-        manifest,
-        fixture,
-        undefined,
-        defaultOptions
-      )
-    ).toThrowError("No React Native profile could satisfy all dependencies");
-
-    expect(consoleErrorSpy).toBeCalledWith(
-      "error",
-      expect.stringContaining(
-        "No React Native profile could satisfy all dependencies"
-      )
-    );
-    expect(consoleWarnSpy).not.toBeCalled();
-  });
-
-  test("does not throw if no profiles can satisfy requirement of dependencies in loose mode", () => {
-    const [fixture, manifest] = useFixture("no-profile-satisfying-deps");
-    expect(() =>
-      getRequirements("^0.64", "app", manifest, fixture, undefined, {
-        loose: true,
-      })
-    ).not.toThrow();
-
-    expect(consoleErrorSpy).not.toBeCalled();
-    expect(consoleWarnSpy).toBeCalledWith(
-      "warn",
-      expect.stringContaining(
-        "No React Native profile could satisfy all dependencies"
-      )
     );
   });
 });
