+++ packages/align-deps/test/vigilant.test.ts
@@ -1,110 +1,67 @@
 import { parseProfilesString } from "../src/profiles";
 import {
   buildManifestProfile,
-  buildProfileFromConfig,
+  checkPackageManifestUnconfigured,
   inspect,
-  makeVigilantCommand,
-} from "../src/vigilant";
+} from "../src/commands/vigilant";
+import type { AlignDepsConfig } from "../src/types";
 
 jest.mock("fs");
+
+function makeConfig(
+  requirements: AlignDepsConfig["alignDeps"]["requirements"],
+  manifest: AlignDepsConfig["manifest"] = {
+    name: "@rnx-kit/align-deps",
+    version: "1.0.0-test",
+  }
+): AlignDepsConfig {
+  return {
+    kitType: "library" as const,
+    alignDeps: {
+      presets: ["microsoft/react-native"],
+      requirements,
+      capabilities: [],
+    },
+    manifest,
+  };
+}
 
 describe("buildManifestProfile()", () => {
   const testVersion = "1.0.0-test";
 
   test("builds a package manifest for a single profile version", () => {
-    const profiles = parseProfilesString("0.64", undefined);
-    const profile = buildManifestProfile(profiles);
+    const profile = buildManifestProfile(
+      "package.json",
+      makeConfig(["react-native@0.70"])
+    );
     profile.version = testVersion;
     expect(profile).toMatchSnapshot();
   });
 
   test("builds a package manifest for multiple profile versions", () => {
-    const profiles = parseProfilesString("0.64,0.63", undefined);
-    const profile = buildManifestProfile(profiles);
+    const profile = buildManifestProfile(
+      "package.json",
+      makeConfig({
+        development: ["react-native@0.70"],
+        production: ["react-native@0.69 || 0.70"],
+      })
+    );
     profile.version = testVersion;
     expect(profile).toMatchSnapshot();
   });
 
   test("includes devOnly packages under `dependencies`", () => {
-    const profiles = parseProfilesString("0.64", undefined);
     const { dependencies, devDependencies, peerDependencies } =
-      buildManifestProfile(profiles);
+      buildManifestProfile("package.json", makeConfig(["react-native@0.70"]));
 
     expect("react-native-test-app" in dependencies).toBe(true);
     expect("react-native-test-app" in peerDependencies).toBe(false);
     expect("react-native-test-app" in devDependencies).toBe(true);
   });
 
-  test("includes custom profiles", () => {
-    const skynet = { name: "skynet", version: "1.0.0" };
-    jest.mock(
-      "vigilant-custom-profiles",
-      () => ({ "0.64": { [skynet.name]: skynet } }),
-      { virtual: true }
-    );
-
-    const profiles = parseProfilesString("0.64", "vigilant-custom-profiles");
-    const { dependencies, devDependencies, peerDependencies } =
-      buildManifestProfile(profiles);
-
-    expect(skynet.name in dependencies).toBe(true);
-    expect(skynet.name in peerDependencies).toBe(true);
-    expect(skynet.name in devDependencies).toBe(true);
-  });
-
   test("throws when no profiles match the requested versions", () => {
     expect(() => parseProfilesString("0.59", undefined)).toThrow();
     expect(() => parseProfilesString("0.59,0.64", undefined)).toThrow();
-  });
-});
-
-describe("buildProfileFromConfig()", () => {
-  const profiles = parseProfilesString("0.64", undefined);
-  const defaultProfile = buildManifestProfile(profiles);
-
-  test("returns default profile if there is no config", () => {
-    expect(buildProfileFromConfig(0, defaultProfile)).toBe(defaultProfile);
-  });
-
-  test("filters out managed capabilities", () => {
-    const dependencies = [
-      "dependencies",
-      "peerDependencies",
-      "devDependencies",
-    ] as const;
-    const config = {
-      kitType: "library" as const,
-      reactNativeVersion: "0.64",
-      reactNativeDevVersion: "0.64",
-      capabilities: [],
-      manifest: {
-        name: "@rnx-kit/align-deps",
-        version: "1.0.0",
-        dependencies: {
-          "react-native": "^0.64.0",
-        },
-      },
-    };
-    const profile = buildProfileFromConfig(config, defaultProfile);
-    dependencies.forEach((section) => {
-      expect(Object.keys(profile[section])).toContain("react");
-      expect(Object.keys(profile[section])).toContain("react-native");
-    });
-
-    const withCapabilities = buildProfileFromConfig(
-      {
-        ...config,
-        capabilities: ["core-android", "core-ios"],
-      },
-      defaultProfile
-    );
-
-    dependencies.forEach((section) => {
-      expect(Object.keys(withCapabilities[section])).not.toContain("react");
-      expect(Object.keys(withCapabilities[section])).not.toContain(
-        "react-native"
-      );
-    });
   });
 });
 
@@ -259,7 +216,7 @@ describe("inspect()", () => {
   });
 });
 
-describe("makeVigilantCommand()", () => {
+describe("checkPackageManifestUnconfigured()", () => {
   const rnxKitConfig = require("@rnx-kit/config");
   const fs = require("fs");
 
@@ -277,104 +234,93 @@ describe("makeVigilantCommand()", () => {
     jest.clearAllMocks();
   });
 
-  test("returns no command if no versions are specified", () => {
-    expect(
-      makeVigilantCommand({ versions: "", loose: false, write: false })
-    ).toBeUndefined();
-  });
-
   test("returns exit code 0 when there are no violations", () => {
-    fs.__setMockContent({
-      name: "@rnx-kit/align-deps",
-      version: "1.0.0",
-      dependencies: {
-        "react-native": "^0.63.2",
-      },
-    });
-
     let didWrite = false;
     fs.__setMockFileWriter(() => {
       didWrite = true;
     });
 
-    const result = makeVigilantCommand({
-      versions: "0.63",
-      loose: false,
-      write: false,
-    })?.("package.json");
+    const result = checkPackageManifestUnconfigured(
+      "package.json",
+      { loose: false, write: false },
+      makeConfig(["react-native@0.70"], {
+        name: "@rnx-kit/align-deps",
+        version: "1.0.0",
+        dependencies: {
+          "react-native": "^0.70.0",
+        },
+      })
+    );
     expect(result).toBe("success");
     expect(didWrite).toBe(false);
     expect(consoleErrorSpy).not.toBeCalled();
   });
 
   test("returns non-zero exit code when there are violations", () => {
-    fs.__setMockContent({
-      name: "@rnx-kit/align-deps",
-      version: "1.0.0",
-      dependencies: {
-        "react-native": "0.63.2",
-      },
-    });
-
     let didWrite = false;
     fs.__setMockFileWriter(() => {
       didWrite = true;
     });
 
-    const result = makeVigilantCommand({
-      versions: "0.63",
-      loose: false,
-      write: false,
-    })?.("package.json");
+    const result = checkPackageManifestUnconfigured(
+      "package.json",
+      { loose: false, write: false },
+      makeConfig(["react-native@0.70"], {
+        name: "@rnx-kit/align-deps",
+        version: "1.0.0",
+        dependencies: {
+          "react-native": "1000.0.0",
+        },
+      })
+    );
     expect(result).not.toBe("success");
     expect(didWrite).toBe(false);
     expect(consoleErrorSpy).toBeCalledTimes(1);
   });
 
   test("returns exit code 0 when writing", () => {
-    fs.__setMockContent({
-      name: "@rnx-kit/align-deps",
-      version: "1.0.0",
-      dependencies: {
-        "react-native": "0.63.2",
-      },
-    });
-
     let didWrite = false;
     fs.__setMockFileWriter(() => {
       didWrite = true;
     });
 
-    const result = makeVigilantCommand({
-      versions: "0.63",
-      loose: false,
-      write: true,
-    })?.("package.json");
+    const result = checkPackageManifestUnconfigured(
+      "package.json",
+      { loose: false, write: true },
+      makeConfig(["react-native@0.70"], {
+        name: "@rnx-kit/align-deps",
+        version: "1.0.0",
+        dependencies: {
+          "react-native": "1000.0.0",
+        },
+      })
+    );
     expect(result).toBe("success");
     expect(didWrite).toBe(true);
     expect(consoleErrorSpy).not.toBeCalled();
   });
 
   test("excludes specified packages", () => {
-    fs.__setMockContent({
-      name: "@rnx-kit/align-deps",
-      version: "1.0.0",
-      dependencies: {
-        "react-native": "0.59.10",
-      },
-    });
-
     let didWrite = false;
     fs.__setMockFileWriter(() => {
       didWrite = true;
     });
 
-    const result = makeVigilantCommand({
-      versions: "0.63",
-      write: false,
-      excludePackages: "@rnx-kit/align-deps",
-      loose: false,
-    })?.("package.json");
+    const result = checkPackageManifestUnconfigured(
+      "package.json",
+      {
+        write: false,
+        excludePackages: ["@rnx-kit/align-deps"],
+        loose: false,
+      },
+      makeConfig(["react-native@0.70"], {
+        name: "@rnx-kit/align-deps",
+        version: "1.0.0",
+        dependencies: {
+          "react-native": "1000.0.0",
+        },
+      })
+    );
     expect(result).toBe("success");
     expect(didWrite).toBe(false);
     expect(consoleErrorSpy).not.toBeCalled();
@@ -383,7 +329,18 @@ describe("makeVigilantCommand()", () => {
   test("uses package-specific custom profiles", () => {
     const fixture = `${__dirname}/__fixtures__/config-custom-profiles-only`;
     const kitConfig = {
-      customProfiles: `${fixture}/packageSpecificProfiles.js`,
+      kitType: "library" as const,
+      alignDeps: {
+        presets: [
+          "microsoft/react-native",
+          `${fixture}/packageSpecificProfiles.js`,
+        ],
+        requirements: {
+          development: ["react-native@0.64"],
+          production: ["react-native@0.64 || 0.65"],
+        },
+        capabilities: [],
+      },
     };
     const inputManifest = {
       name: "@rnx-kit/align-deps",
@@ -400,18 +357,18 @@ describe("makeVigilantCommand()", () => {
     };
 
     rnxKitConfig.__setMockConfig(kitConfig);
-    fs.__setMockContent(inputManifest);
 
     let manifest = undefined;
     fs.__setMockFileWriter((_, content) => {
       manifest = JSON.parse(content);
     });
 
-    const result = makeVigilantCommand({
-      versions: "0.64,0.65",
-      loose: false,
-      write: true,
-    })?.("package.json");
+    const result = checkPackageManifestUnconfigured(
+      "package.json",
+      { loose: false, write: true },
+      { ...kitConfig, manifest: inputManifest }
+    );
+
     expect(result).toBe("success");
     expect(consoleErrorSpy).not.toBeCalled();
     expect(manifest).toEqual({
@@ -422,55 +379,7 @@ describe("makeVigilantCommand()", () => {
       },
       peerDependencies: {
         react: "17.0.2",
-        "react-native": "0.64.3 || 0.65.2 || ^0.64.2 || ^0.65.0",
-      },
-    });
-  });
-
-  test("prefers package-specific React Native versions", () => {
-    const fixture = `${__dirname}/__fixtures__/config-custom-profiles-only`;
-    const kitConfig = {
-      reactNativeVersion: "0.64",
-      customProfiles: `${fixture}/packageSpecificProfiles.js`,
-    };
-    const inputManifest = {
-      name: "@rnx-kit/align-deps",
-      version: "1.0.0",
-      peerDependencies: {
-        react: "17.0.1",
-        "react-native": "0.64.0",
-      },
-      devDependencies: {
-        react: "17.0.1",
-        "react-native": "0.64.0",
-      },
-      "rnx-kit": kitConfig,
-    };
-
-    rnxKitConfig.__setMockConfig(kitConfig);
-    fs.__setMockContent(inputManifest);
-
-    let manifest = undefined;
-    fs.__setMockFileWriter((_, content) => {
-      manifest = JSON.parse(content);
-    });
-
-    const result = makeVigilantCommand({
-      versions: "0.64,0.65",
-      loose: false,
-      write: true,
-    })?.("package.json");
-    expect(result).toBe("success");
-    expect(consoleErrorSpy).not.toBeCalled();
-    expect(manifest).toEqual({
-      ...inputManifest,
-      devDependencies: {
-        react: "17.0.2",
-        "react-native": "0.64.3",
-      },
-      peerDependencies: {
-        react: "17.0.2",
-        "react-native": "0.64.3 || ^0.64.2",
+        "react-native": "0.64.3 || 0.65.2",
       },
     });
   });
